Add a stub method for our endgoal test.
This test passing represents our desired end state. It's going to fail
for a good while and that's expected and okay. When I have defined goals
like this I really like to lay them out early since it provides a focus:
"What do I need to do next to get closer to this passing?"

For now the implementation method just writes out the input data.
That won't do but it will for now.