Add files specific to my use case.
The specific case I'm building this for has metadata in the filenames
that needs to be extracted and inserted into the final table output.
Since this is going to be highly specific I'm going to keep it separate
so it can be removed or altered by any future users.

The initial method here just pulls the site from the filename since it's
recorded in a more consistent format in the filename than in the sunscan
itself..