Pull year from filename.
The date recorded in the sunscan is unreliable due to the realtime clock
not being reset after changing the batteries. In the sample dataset, the
exact date isn't needed and just the year is enough.

Again the 2011 data is named with a different pattern and so needs to be
treated specially.