Parse tabular data and record as output.
The parse method is getting pretty heavy, but it's also the bulk of what
we're doing and I'm not sure it makes sense to decompose it yet.

The sundata files are divided into frontmatter data and tabular data.
The only indication that we're entering the Tabular section is the table
header so we use that as a key to the special variable
`arrived_at_tabular_data` a mouthful but accurate.

There's a few frontmatter checks added already, but they're mostly
skipping lines that don't contain anything worth recording.
Later we'll add things like the date and other metadata up there.

For now we're mostly focused on pulling out the tabular data from each
table entry.