Correct parsing of multi-key lines.
There were two issues to resolve here:

The regexp was matching the tab characters between each field which it
was doing since there was an arbitrary number of them. I replaced the
regular expression with one that matches any characters that aren't tabs
or colons, and does it's best to strip whitespace around fields.

This worked better, but it still meant that only the first intance of
the pattern was read. This took a bit of digging and collaborating with
another regexp obsessed friend. We were using Rubular.com to collaborate
but Rubular returns _all_ matches for a particular string whereas
String#match and Regexp#match return only the first. We needed
String#scan. Now it works!